touch: goodix: fix suspend resume sequence for dual touch case
Touch resume and suspend is done only when core_module_prob_sate
global variable is set to probed state. In case of dual touch,
if secondary probe fails, core_module_prob_sate is updated to
probe failed state. Due to this, suspend or resume calls for
primary touch returns early even though primary touch probe is
successful. This leads to primary touch remain always on.

To fix this, add a new ready variable in touch data structure
and perform suspend or resume operations when ready flag is
set to true.

Change-Id: I7b93a5b3135736caa3e2b5aaf6ff7d8adf4dee9d
Signed-off-by: Ritesh Kumar <quic_riteshk@quicinc.com>