bolero: rx-macro: notify swr wcd bus clk frequency from hw params
Notify bus clk frequency change for swr from hw params function.

Earlier it was notified only as part of RX_MCLK widget power
up function, But this does not help to notify when ever two Rx use case
running in parallel and device switch happens, above widget is not
powered up again as alreay it is powered up for one of the use case.

So it is good to have as part of hw params function which gets
called always for every use case, which helps in taking the
updated clk frequency for native playback

Change-Id: I4516947afe6e7656ccf7f8b6ee1badbaef96cfce