disp: msm: change memory allocation to kvzalloc
Change memory allocation from kzalloc to kvzalloc for
rotator device. kzalloc do not gurantee the physical
contiguous allocations for bigger memory size.
kvzalloc will try to allocate memory with kmalloc and
if the allocation fails it will be retried with vmalloc.
kvzalloc may return memory that is not physically contiguous.

Change-Id: I09c7d2c30b7c61e4274762f28e4dfe1a1fa4a612
Signed-off-by: Raghevendra Ambadas <quic_c_rambad@quicinc.com>
Signed-off-by: Nirmal Abraham <quic_c_nabrah@quicinc.com>